# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nagibe Produção API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /team:
    get:
      operationId: listTeamMembers
      tags: [team]
      summary: List all team members
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: createTeamMember
      tags: [team]
      summary: Create a team member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamMemberBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /team/{id}:
    get:
      operationId: getTeamMember
      tags: [team]
      summary: Get a team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
        "404":
          description: Not found
    patch:
      operationId: updateTeamMember
      tags: [team]
      summary: Update a team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamMemberBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
    delete:
      operationId: deleteTeamMember
      tags: [team]
      summary: Delete a team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ──────── Equipment ────────
  /equipment:
    get:
      operationId: listEquipment
      tags: [equipment]
      summary: List all equipment
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of equipment
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Equipment"
    post:
      operationId: createEquipment
      tags: [equipment]
      summary: Create equipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEquipmentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
  /equipment/{id}:
    get:
      operationId: getEquipment
      tags: [equipment]
      summary: Get equipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
        "404":
          description: Not found
    patch:
      operationId: updateEquipment
      tags: [equipment]
      summary: Update equipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEquipmentBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
    delete:
      operationId: deleteEquipment
      tags: [equipment]
      summary: Delete equipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ──────── Shoots ────────
...
components:
  schemas:
...
    TeamMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        primaryRole:
          type: string
        secondaryRole:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, primaryRole, status, createdAt, updatedAt]
    CreateTeamMemberBody:
      type: object
      properties:
        name:
          type: string
        primaryRole:
          type: string
        secondaryRole:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
          default: active
      required: [name, primaryRole]
    UpdateTeamMemberBody:
      type: object
      properties:
        name:
          type: string
        primaryRole:
          type: string
        secondaryRole:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
    Equipment:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        internalCode:
          type: ["string", "null"]
        totalQuantity:
          type: integer
        availableQuantity:
          type: integer
        status:
          type: string
          enum: [available, reserved, in_use, maintenance, damaged, pending_return]
        storageLocation:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, category, totalQuantity, availableQuantity, status, active, createdAt, updatedAt]
    CreateEquipmentBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        internalCode:
          type: ["string", "null"]
        totalQuantity:
          type: integer
          default: 1
        availableQuantity:
          type: integer
          default: 1
        status:
          type: string
          enum: [available, reserved, in_use, maintenance, damaged, pending_return]
          default: available
        storageLocation:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        active:
          type: boolean
          default: true
      required: [name, category]
    UpdateEquipmentBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        internalCode:
          type: ["string", "null"]
        totalQuantity:
          type: integer
        availableQuantity:
          type: integer
        status:
          type: string
          enum: [available, reserved, in_use, maintenance, damaged, pending_return]
        storageLocation:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        active:
          type: boolean
...