# lib/api-zod/src/generated/types/updateEquipmentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nagibe Produção API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateEquipmentBodyStatus } from "./updateEquipmentBodyStatus";

export interface UpdateEquipmentBody {
  name?: string;
  category?: string;
  /** @nullable */
  internalCode?: string | null;
  totalQuantity?: number;
  availableQuantity?: number;
  status?: UpdateEquipmentBodyStatus;
  /** @nullable */
  storageLocation?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  active?: boolean;
}